fix: fix get camera and call log (#1)
The error "self.deviceId is None" when calling `getLastPicture` is due to the fact that by placing the token retrieval process inside another thread, the `getLastPicture` method is called first. It's not actually necessary to obtain this token in another thread (instead of keeping the notification listener in a separate thread).

I've moved the token retrieval process outside the main thread, making it synchronous (since it's needed for other methods), and left the notification listener in the main thread.

There's also a fix in the `getLastPicture` method itself to set a default value of "max", because when `None` is passed if there are no call records, it returns an error.